Reorganize geolocation's packaging structure
