+++ app/src/androidTest/java/com/github/onedirection/geolocation/NominatimGeocodingTest.java
@@ -4,6 +4,7 @@ import android.content.Context;
 
 import androidx.test.core.app.ApplicationProvider;
 
+import com.github.onedirection.geolocation.geocoding.NominatimGeocoding;
 import com.github.onedirection.utils.Monads;
 
 import org.junit.Test;
