+++ app/src/main/java/com/github/onedirection/events/EventCreator.java
@@ -11,7 +11,7 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.test.espresso.IdlingResource;
 
 import com.github.onedirection.R;
-import com.github.onedirection.geolocation.DeviceLocationProviderActivity;
+import com.github.onedirection.geolocation.location.DeviceLocationProviderActivity;
 
 import java.time.LocalDate;
