+++ app/src/main/java/com/github/onedirection/geolocation/location/AbstractDeviceLocationProvider.java
@@ -1,4 +1,4 @@
-package com.github.onedirection.geolocation;
+package com.github.onedirection.geolocation.location;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -7,6 +7,7 @@ import android.os.Looper;
 
 import androidx.annotation.NonNull;
 
+import com.github.onedirection.geolocation.Coordinates;
 import com.github.onedirection.utils.Monads;
 import com.github.onedirection.utils.ObserverPattern;
 import com.google.android.gms.location.FusedLocationProviderClient;
