+++ app/src/main/java/com/github/onedirection/navigation/NavigationActivity.java
@@ -19,7 +19,6 @@ import androidx.navigation.ui.NavigationUI;
 import com.github.onedirection.R;
 import com.github.onedirection.authentication.FirebaseAuthentication;
 import com.github.onedirection.events.EventCreator;
-import com.github.onedirection.geolocation.DeviceLocationProviderActivity;
 import com.google.android.material.navigation.NavigationView;
 
 public class NavigationActivity extends AppCompatActivity {
